Crate ocl_core

Thin wrappers for the OpenCL FFI functions and types.

Allows access to OpenCL FFI functions with a minimal layer of abstraction, providing both safety and convenience. The ocl crate contains higher level and easier to use interfaces to the functionality contained within.

Even Lower Level: cl-sys

If there's still something missing or for some reason you need direct FFI access, use the functions in the cl-sys module. The pointers used by cl-sys functions can be wrapped in ocl-core wrappers (ocl_core::PlatformId, ocl_core::Context, etc.) and passed to ocl-core module functions. Likewise the other way around (using, for example: [EventRaw::as_ptr]).

Performance

Performance between all three interface layers, cl-sys, ocl-core, and the 'standard' (ocl) types, is identical or virtually identical (if not, please file an issue).

Safety

At the time of writing, some functions still may break Rust's usual safety promises and have not been 100% comprehensively evaluated and tested. Please file an issue if you discover something!

Length vs Size

No, not that...

Quantifiers passed to functions in the OpenCL API are generally expressed in bytes. Units passed to functions in this library are expected to be bytes / sizeof(T) (corresponding with units returned by the ubiquitous .len() method). The suffix '_size' or '_bytes' is generally used when a parameter deviates from this convention.

Version Control

The version control system is in place to ensure that you don't call functions that your hardware/driver does not support.

Functions in this crate with the [Version Controlled: OpenCL {...}+] tag in the description require an additional parameter, device_version or device_versions: a parsed result (or slice of results) of DeviceInfo::Version. This is a runtime check to ensure that the device supports the function being called. Calling a function which a particular device does not support will likely cause a segmentation fault and possibly data corruption.

Saving the OpenclVersion returned from device_version() for your device(es) at the start of your program and passing it each time you call a version controlled function is the fastest and safest method (see the ocl library for an example). The cost of this check is little more than a single if statement.

Passing None for device_version will cause an automated version check which has a small cost (calling info function, parsing the version number etc.) but is a safe option if you are not sure what to do.

Passing the result of a call to OpenclVersion::max() or passing a fake version will bypass any safety checks and has all of the risks described above. Only do this if you're absolutely sure you know what you're doing and are not concerned about segfaults and data integrity.

More Documentation

As most of the functions here are minimally documented, please refer to the official OpenCL documentation linked below. Although there isn't a precise 1:1 parameter mapping between the core and original functions, it's close enough (modulo the size/len difference discussed above) to help sort out any questions you may have until a more thorough documentation pass can be made. View the source code in src/types/functions.rs for more mapping details.

'OpenCL' 1.2 SDK Reference: https://www.khronos.org/registry/cl/sdk/1.2/docs/man/xhtml/

Reexports

pub extern crate cl_sys as ffi;

pub use self::types::abs::{ClEventPtrNew, ClEventRef, ClPlatformIdPtr, ClDeviceIdPtr, EventRefWrapper, PlatformId, DeviceId, Context, CommandQueue, Mem, Program, Kernel, Event, EventList, Sampler, ClWaitList, ClVersions};

pub use self::types::structs::{OpenclVersion, ContextProperties, ImageFormat, ImageDescriptor, BufferRegion, ContextPropertyValue};

pub use self::types::enums::{KernelArg, PlatformInfoResult, DeviceInfoResult, ContextInfoResult, CommandQueueInfoResult, MemInfoResult, ImageInfoResult, SamplerInfoResult, ProgramInfoResult, ProgramBuildInfoResult, KernelInfoResult, KernelArgInfoResult, KernelWorkGroupInfoResult, EventInfoResult, ProfilingInfoResult};

pub use self::types::vectors::{ClChar2, ClChar3, ClChar4, ClChar8, ClChar16, ClUchar2, ClUchar3, ClUchar4, ClUchar8, ClUchar16, ClShort2, ClShort3, ClShort4, ClShort8, ClShort16, ClUshort2, ClUshort3, ClUshort4, ClUshort8, ClUshort16, ClInt2, ClInt3, ClInt4, ClInt8, ClInt16, ClUint2, ClUint3, ClUint4, ClUint8, ClUint16, ClLong1, ClLong2, ClLong3, ClLong4, ClLong8, ClLong16, ClUlong1, ClUlong2, ClUlong3, ClUlong4, ClUlong8, ClUlong16, ClFloat2, ClFloat3, ClFloat4, ClFloat8, ClFloat16, ClDouble2, ClDouble3, ClDouble4, ClDouble8, ClDouble16};

Modules

error
Standard error type for ocl.

types

util
Utility and debugging functions.

Structs

CommandQueueProperties
cl_command_queue_properties - bitfield

DeviceAffinityDomain
cl_device_affinity_domain

DeviceExecCapabilities
cl_device_exec_capabilities - bitfield

DeviceFpConfig
cl_device_fp_config - bitfield

DeviceType
cl_device_type - bitfield

KernelArgTypeQualifier
cl_kernel_arg_type_qualifer

MapFlags
cl_map_flags - bitfield

MemFlags
cl_mem_flags - bitfield

MemMigrationFlags
cl_mem_migration_flags - bitfield

ProgramBinaryType
cl_program_binary_type

Enums

AddressingMode
cl_addressing_mode

BufferCreateType
cl_buffer_create_type

Cbool
cl_bool

ChannelType
cl_channel_type

ClGlObjectType

CommandExecutionStatus
command execution status

CommandQueueInfo
cl_command_queue_info

CommandType
cl_command_type

ContextInfo
cl_context_info

ContextInfoOrPropertiesPointerType
cl_context_info + cl_context_properties

ContextProperty
cl_context_info + cl_context_properties

DeviceInfo
cl_device_info

DeviceLocalMemType
cl_device_local_mem_type

DeviceMemCacheType
cl_mem_cache_type

DevicePartitionProperty
[INCOMPLETE] cl_device_partition_property

EventInfo
cl_event_info

FilterMode
cl_filter_mode

GlTextureTarget
specify the texture target type

ImageChannelDataType
Describes the size of the channel data type. The number of bits per element determined by the image_channel_data_type and image_channel_order must be a power of two. The list of supported values is described in the table below. (from SDK)

ImageChannelOrder
Specifies the number of channels and the channel layout i.e. the memory layout in which channels are stored in the image. Valid values are described in the table below. (from SDK)

ImageInfo
cl_image_info

KernelArgAccessQualifier
cl_kernel_arg_access_qualifier

KernelArgAddressQualifier
cl_kernel_arg_address_qualifier

KernelArgInfo
cl_kernel_arg_info

KernelInfo
cl_kernel_info

KernelWorkGroupInfo
cl_kernel_work_group_info

MemInfo
cl_mem_info

MemObjectType
cl_mem_object_type

PlatformInfo
cl_platform_info

Polling
cl_bool: Polling

ProfilingInfo
cl_profiling_info

ProgramBuildInfo
cl_program_build_info

ProgramBuildStatus
cl_build_status

ProgramInfo
cl_program_info

SamplerInfo
cl_sampler_info

Status
TODO: MOVE ME AND LEAVE CONSTS AS THEY WERE. The status of an OpenCL API call. Used for returning success/error codes.

Constants

DEVICES_MAX

DEVICE_AFFINITY_DOMAIN_L1_CACHE

DEVICE_AFFINITY_DOMAIN_L2_CACHE

DEVICE_AFFINITY_DOMAIN_L3_CACHE

DEVICE_AFFINITY_DOMAIN_L4_CACHE

DEVICE_AFFINITY_DOMAIN_NEXT_PARTITIONABLE

DEVICE_AFFINITY_DOMAIN_NUMA

DEVICE_TYPE_ACCELERATOR

DEVICE_TYPE_ALL

DEVICE_TYPE_CPU

DEVICE_TYPE_CUSTOM

DEVICE_TYPE_DEFAULT

DEVICE_TYPE_GPU

EXEC_KERNEL

EXEC_NATIVE_KERNEL

FP_CORRECTLY_ROUNDED_DIVIDE_SQRT

FP_DENORM

FP_FMA

FP_INF_NAN

FP_ROUND_TO_INF

FP_ROUND_TO_NEAREST

FP_ROUND_TO_ZERO

FP_SOFT_FLOAT

KERNEL_ARG_TYPE_CONST

KERNEL_ARG_TYPE_NONE

KERNEL_ARG_TYPE_RESTRICT

KERNEL_ARG_TYPE_VOLATILE

MAP_READ

MAP_WRITE

MAP_WRITE_INVALIDATE_REGION

MEM_ALLOC_HOST_PTR

MEM_COPY_HOST_PTR

MEM_HOST_NO_ACCESS

MEM_HOST_READ_ONLY

MEM_HOST_WRITE_ONLY

MEM_READ_ONLY

MEM_READ_WRITE

MEM_USE_HOST_PTR

MEM_WRITE_ONLY

MIGRATE_MEM_OBJECT_CONTENT_UNDEFINED

MIGRATE_MEM_OBJECT_HOST

PROGRAM_BINARY_TYPE_COMPILED_OBJECT

PROGRAM_BINARY_TYPE_EXECUTABLE

PROGRAM_BINARY_TYPE_LIBRARY

PROGRAM_BINARY_TYPE_NONE

QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE

QUEUE_PROFILING_ENABLE

Traits

OclPrm
A type usable within OpenCL kernels.

OclScl
A scalar type usable within OpenCL kernels.

OclVec
A vector type usable within OpenCL kernels.

Functions

build_program
Builds a program.

compile_program
[UNIMPLEMENTED]

create_buffer
Returns a new buffer pointer with size (bytes): len * sizeof(T).

create_build_program
Creates, builds, and returns a new program pointer from src_strings.

create_command_queue
Returns a new command queue pointer.

create_context
Creates a new context pointer valid for all devices in device_ids.

create_context_from_type
Creates a new context pointer for all devices of a specific type.

create_from_gl_buffer
[UNTESTED] Return a buffer pointer from a OpenGL buffer object.

create_from_gl_renderbuffer
[UNTESTED] Return a renderbuffer pointer from a OpenGL renderbuffer object.

create_from_gl_texture
[UNTESTED] Return a texture2D pointer from a OpenGL texture2D object.

create_from_gl_texture_2d
[UNTESTED] [DEPRICATED] Return a texture2D pointer from a OpenGL texture2D object.

create_from_gl_texture_3d
[UNTESTED] [DEPRICATED] Return a texture3D pointer from a OpenGL texture3D object.

create_image
Returns a new image (mem) pointer.

create_kernel
Returns a new kernel pointer.

create_kernels_in_program
[UNIMPLEMENTED]

create_program_with_binary
[UNTESTED] Creates a program object for a context, and loads the binary bits specified by binary into the program object.

create_program_with_built_in_kernels
[UNIMPLEMENTED]

create_program_with_source
Creates a new program.

create_sampler
Creates and returns a new sampler object.

create_sub_buffer
[UNTESTED] Creates a new buffer object (referred to as a sub-buffer object) from an existing buffer object.

create_sub_devices
[UNIMPLEMENTED]

create_user_event
[UNTESTED] Creates an event not already associated with any command.

default_device_type
Returns the default device type bitflags as specified by environment variable or else DEVICE_TYPE_ALL.

default_platform
Returns the default or first platform.

default_platform_idx
Returns the default platform if set by an environment variable or config file.

device_versions
Returns a list of versions for devices.

enqueue_acquire_gl_buffer
[UNTESTED] Enqueue acquire OpenCL memory objects that have been created from OpenGL objects.

enqueue_barrier_with_wait_list
[UNTESTED] A synchronization point that enqueues a barrier operation.

enqueue_copy_buffer
[UNTESTED] Copies the contents of one buffer to another.

enqueue_copy_buffer_rect
Enqueues a command to copy a rectangular region from a buffer object to another buffer object.

enqueue_copy_buffer_to_image
[UNTESTED] Enqueues a command to copy a buffer object to an image object.

enqueue_copy_image
Enqueues a command to copy image objects.

enqueue_copy_image_to_buffer
[UNTESTED] Enqueues a command to copy an image object to a buffer object.

enqueue_fill_buffer
Enqueues a command to fill a buffer object with a pattern of a given pattern size.

enqueue_fill_image
[UNTESTED] Enqueues a command to fill an image object with a specified color.

enqueue_kernel
Enqueues a command to execute a kernel on a device.

enqueue_map_buffer
[UNTESTED] Enqueues a command to map a region of the buffer object given by buffer into the host address space and returns a pointer to this mapped region.

enqueue_map_image
[UNTESTED] Enqueues a command to map a region of the image object given by image into the host address space and returns a pointer to this mapped region.

enqueue_marker_with_wait_list
[UNTESTED] Enqueues a marker command which waits for either a list of events to complete, or all previously enqueued commands to complete.

enqueue_migrate_mem_objects
[UNTESTED] Enqueues a command to indicate which device a set of memory objects should be associated with.

enqueue_native_kernel
[UNIMPLEMENTED]

enqueue_read_buffer
Enqueues a read from device memory referred to by buffer to device memory, data.

enqueue_read_buffer_rect
Enqueues a command to read from a rectangular region from a buffer object to host memory.

enqueue_read_image
Reads an image from device to host memory.

enqueue_release_gl_buffer
[UNTESTED] Enqueue release OpenCL memory objects that have been created from OpenGL objects.

enqueue_task
[UNTESTED] Enqueues a command to execute a kernel on a device.

enqueue_unmap_mem_object
[UNTESTED] Enqueues a command to unmap a previously mapped region of a memory object.

enqueue_write_buffer
Enqueues a write from host memory, data, to device memory referred to by buffer.

enqueue_write_buffer_rect
Enqueues a command to write from a rectangular region from host memory to a buffer object.

enqueue_write_image
Enqueues a command to write to an image or image array object from host memory.

finish
Waits for a queue to finish.

flush
[UNTESTED] Flushes a command queue.

get_command_queue_info
Returns information about a command queue

get_context_info
Returns various kinds of context information.

get_device_ids
Returns a list of available devices for a particular platform.

get_device_info
Returns information about a device.

get_event_info
Get event info.

get_event_profiling_info
Get event profiling info (for debugging / benchmarking).

get_event_status
Returns the status of event.

get_extension_function_address_for_platform
[UNTESTED] Returns the address of the extension function named by func_name for a given platform.

get_image_info
Get image info.

get_kernel_arg_info
Get kernel arg info.

get_kernel_info
Get kernel info.

get_kernel_work_group_info
Get kernel work group info.

get_mem_object_info
Get mem object info.

get_platform_ids
Returns a list of available platforms as 'core' objects.

get_platform_info
Returns platform information of the requested type.

get_program_build_info
Get program build info.

get_program_info
Get program info.

get_sampler_info
Returns information about the sampler object.

get_supported_image_formats
Returns a list of supported image formats.

link_program
[UNIMPLEMENTED]

program_build_err
If the program pointed to by cl_program for any of the devices listed in device_ids has a build log of any length, it will be returned as an errcode result.

release_command_queue
Decrements the reference count of a command queue.

release_context
Decrements reference count of a context.

release_device
Decrements the reference count of a device.

release_event
Decrements an event's reference counter.

release_kernel
Decrements a kernel reference counter.

release_mem_object
Decrements the reference counter of a mem object.

release_program
Decrements a program reference counter.

release_sampler
Decrements a sampler reference counter.

retain_command_queue
Increments the reference count of a command queue.

retain_context
Increments the reference count of a context.

retain_device
Increments the reference count of a device.

retain_event
Increments an event's reference counter.

retain_kernel
Increments a kernel reference counter.

retain_mem_object
Increments the reference counter of a mem object.

retain_program
Increments a program reference counter.

retain_sampler
Increments a sampler reference counter.

set_event_callback
Sets a callback function which is called as soon as the callback_trigger status is reached.

set_kernel_arg
Sets the argument value for a specific argument of a kernel.

set_mem_object_destructor_callback
[UNIMPLEMENTED]

set_user_event_status
[UNTESTED] Updates a user events status.

verify_context
Verifies that the context is in fact a context object pointer.

wait_for_event
Blocks until an event is complete.

wait_for_events
Blocks until the first num_events events in event_list are complete.

Type Definitions

BuildProgramCallbackFn

CreateContextCallbackFn

EventCallbackFn

UserDataPtr