Function ocl_core::_complete_user_event

pub extern fn _complete_user_event(
    src_event_ptr: cl_event, 
    event_status: i32, 
    user_data: *mut c_void
)

If event_status is CommandExecutionStatus::Complete, the cl_event pointed to by user_data will be set to the same.

user_data must be a cl_event which has not yet had its destructor run (::release_event).

src_event_ptr is not used and does not need anything special done with its destructor (it will already have been managed by the call to ::set_event_callback.