Function ocl_core::enqueue_fill_image

pub fn enqueue_fill_image<T>(command_queue: &CommandQueue,
                             image: &Mem,
                             color: &[T],
                             origin: [usize; 3],
                             region: [usize; 3],
                             wait_list: Option<&ClWaitList>,
                             new_event: Option<&mut ClEventPtrNew>,
                             device_version: Option<&OpenclVersion>)
                             -> OclResult<()>

[UNTESTED] Enqueues a command to fill an image object with a specified color.

Fill Color (from SDK docs

The fill color. The fill color is a four component RGBA floating-point color value if the image channel data type is not an unnormalized signed and unsigned integer type, is a four component signed integer value if the image channel data type is an unnormalized signed integer type and is a four component unsigned integer value if the image channel data type is an unormalized unsigned integer type. The fill color will be converted to the appropriate image channel format and order associated with image.

TODO: Trait constraints for T. Presumably it should be 32bits? Testing needed.

[Version Controlled: OpenCL 1.2+] See module docs for more info.