Function ocl_core::get_context_info

pub fn get_context_info<C>(
    context: C, 
    request: ContextInfo
) -> ContextInfoResult where
    C: ClContextPtr, 

Returns various kinds of context information.

SDK Reference

Errors

Returns an error result for all the reasons listed in the SDK in addition to an additional error when called with CL_CONTEXT_DEVICES as described in in the verify_context() documentation below.