Function ocl_core::get_extension_function_address_for_platform

pub unsafe fn get_extension_function_address_for_platform(platform: &PlatformId,
                                                          func_name: &str,
                                                          platform_version: Option<&OpenclVersion>)
                                                          -> OclResult<*mut c_void>

[UNTESTED] Returns the address of the extension function named by func_name for a given platform.

The pointer returned should be cast to a function pointer type matching the extension function's definition defined in the appropriate extension specification and header file.

A non-NULL return value does not guarantee that an extension function is actually supported by the platform. The application must also make a corresponding query using ocl::core::get_platform_info(platform_core, CL_PLATFORM_EXTENSIONS, ... ) or ocl::core::get_device_info(device_core, CL_DEVICE_EXTENSIONS, ... ) to determine if an extension is supported by the OpenCL implementation.

[FIXME]: Update enum names above to the wrapped types.

Errors

Returns an error if:

func_name cannot be converted to a CString.

The specified function does not exist for the implementation.

'platform' is not a valid platform.

[Version Controlled: OpenCL 1.2+] See module docs for more info.