Function ocl_core::set_kernel_arg

pub fn set_kernel_arg<T: OclPrm>(
    kernel: &Kernel, 
    arg_index: u32, 
    arg: KernelArg<T>
) -> OclResult<()>

Sets the argument value for a specific argument of a kernel.

SDK Documentation

[FIXME: Remove] kernel_name is for error reporting and is optional but highly recommended.

TODO: Remove name parameter and lookup name with get_kernel_info instead.