Function ocl_core::set_user_event_status

pub fn set_user_event_status<'e, E: ClEventPtrRef<'e>>(
    event: &'e E, 
    execution_status: CommandExecutionStatus
) -> OclCoreResult<()>

Updates a user events status.

Setting status to completion will cause commands waiting upon this event to execute.

Will return an error if the event is not a 'user' event (created with ::create_user_event).

Valid options are (for OpenCL versions 1.1 - 2.1):

CommandExecutionStatus::Complete CommandExecutionStatus::Running CommandExecutionStatus::Submitted CommandExecutionStatus::Queued

To the best of the author's knowledge, the only variant that matters is ::Complete. Everything else is functionally equivalent and is useful only for debugging or profiling purposes (this may change in the future).