Struct ocl_core::types::abs::CommandQueue

#[repr(C)]

pub struct CommandQueue(_);

cl_command_queue

Methods

impl CommandQueue

unsafe fn from_raw_create_ptr(ptr: cl_command_queue) -> CommandQueue

Only call this when passing the original newly created pointer directly from clCreate.... Do not use this to clone or copy.

unsafe fn from_raw_copied_ptr(ptr: cl_command_queue) -> CommandQueue

Only call this when passing a copied pointer such as from an clGet*****Info function.

fn as_ptr(&self) -> cl_command_queue

Returns a pointer, do not store it.

fn device(&self) -> OclResult<DeviceId>

Returns the DeviceId associated with this command queue.

fn context(&self) -> OclResult<Context>

Returns the Context associated with this command queue.

fn context_ptr(&self) -> OclResult<cl_context>

Returns the cl_context associated with this command queue.