Struct ocl_core::types::abs::Context

#[repr(C)]

pub struct Context(_);

cl_context

Methods

impl Context

pub unsafe fn from_raw_create_ptr(ptr: cl_context) -> Context

Only call this when passing the original newly created pointer directly from clCreate.... Do not use this to clone or copy.

pub unsafe fn from_raw_copied_ptr(ptr: cl_context) -> Context

Only call this when passing a copied pointer such as from an clGet*****Info function.

pub fn as_ptr(&self) -> cl_context

Returns a pointer, do not store it.

pub fn devices(&self) -> OclCoreResult<Vec<DeviceId>>

Returns the devices associated with this context.

pub fn platform(&self) -> OclCoreResult<Option<PlatformId>>

Returns the platform associated with this context, if any.

Errors upon the usual OpenCL errors.

Returns None if the context properties do not specify a platform.