Struct ocl_core::types::abs::DeviceId

pub struct DeviceId(_);

cl_device_id

Methods

impl DeviceId

unsafe fn from_fresh_ptr(ptr: cl_device_id) -> DeviceId

Only call this when passing a newly created pointer directly from clCreate.... Do not use this to clone or copy.

unsafe fn from_copied_ptr(ptr: cl_device_id) -> DeviceId

Only call this when passing a copied pointer such as from an clGet*****Info function.

unsafe fn null() -> DeviceId

unsafe fn as_ptr(&self) -> cl_device_id

Returns a pointer.

fn version(&self) -> OclResult<OpenclVersion>

Returns the looked up and parsed OpenCL version for this device.