Struct ocl_core::types::abs::Event

pub struct Event(_);

cl_event

Methods

impl Event

unsafe fn from_fresh_ptr(ptr: cl_event) -> Event

Only call this when passing a newly created pointer directly from clCreate.... Do not use this to clone or copy.

unsafe fn from_cloned_ptr(ptr: cl_event) -> OclResult<Event>

Only use when cloning from a pre-existing and valid cl_event.

unsafe fn null() -> Event

For passage directly to an 'event creation' function (such as enqueue).

unsafe fn as_ptr_ref(&self) -> &cl_event

Returns an immutable reference to a pointer, do not deref and store it unless you will manage its associated reference count carefully.

unsafe fn as_ptr_mut(&mut self) -> &mut cl_event

Returns a mutable reference to a pointer, do not deref then modify or store it unless you will manage its associated reference count carefully.

fn is_valid(&self) -> bool

[FIXME]: ADD VALIDITY CHECK BY CALLING '_INFO' OR SOMETHING: NULL CHECK IS NOT ENOUGH