Struct ocl_core::types::abs::EventList

pub struct EventList { /* fields omitted */ }

List of cl_events.

Methods

impl EventList

fn new() -> EventList

Returns a new, empty, EventList.

fn push(&mut self, event: Event)

Pushes a new event onto the list.

Technically, copies event's contained pointer (a cl_event) then mem::forgets it. This seems preferrable to incrementing the reference count (with functions::retain_event) then letting event drop which just decrements it right back.

fn pop(&mut self) -> Option<OclResult<Event>>

Removes the last event from the list and returns it.

fn allot(&mut self) -> &mut cl_event

Appends a new null element to the end of the list and returns a reference to it.

fn len(&self) -> usize

fn is_empty(&self) -> bool

fn count(&self) -> u32

unsafe fn as_ptr_ref(&self) -> &cl_event

Returns an immutable reference to a pointer, do not deref and store it unless you will manage its associated reference count carefully.

fn get_clone(&self, index: usize) -> Option<OclResult<Event>>

Clones an event by index.

fn last_clone(&self) -> Option<OclResult<Event>>

Clones the last event.

fn clear_completed(&mut self) -> OclResult<()>

Clears each completed event from the list.

TODO: TEST THIS