Struct ocl_core::types::abs::Mem

#[repr(C)]

pub struct Mem(_);

cl_mem

Methods

impl Mem

pub unsafe fn from_raw_create_ptr(ptr: cl_mem) -> Mem

Only call this when passing the original newly created pointer directly from clCreate.... Do not use this to clone or copy.

pub unsafe fn from_raw_copied_ptr(ptr: cl_mem) -> Mem

Only call this when passing a copied pointer such as from an clGet*****Info function.

pub fn as_ptr(&self) -> cl_mem

Returns a pointer, do not store it.