Struct ocl_core::types::abs::Program

#[repr(C)]

pub struct Program(_);

cl_program

Methods

impl Program

unsafe fn from_raw_create_ptr(ptr: cl_program) -> Program

Only call this when passing the original newly created pointer directly from clCreate.... Do not use this to clone or copy.

unsafe fn from_raw_copied_ptr(ptr: cl_program) -> Program

Only call this when passing a copied pointer such as from an clGet*****Info function.

fn as_ptr(&self) -> cl_program

Returns a pointer, do not store it.

fn devices(&self) -> OclResult<Vec<DeviceId>>

Returns the devices associated with this program.