Module ocl_core::types::structs

Rust implementations of various structs used by the OpenCL API.

Structs

ArgVal
A reference to a kernel argument value.

BufferRegion
Defines a buffer region for creating a sub-buffer.

ContextProperties
Context properties list.

ImageDescriptor
An image descriptor use in the creation of Image.

ImageFormat
Image format properties used by Image.

OpenclVersion
Parsed OpenCL version in the layout ({major}, {minor}).

Enums

ContextPropertyValue

ImageFormatParseError

Type Definitions

ImageFormatParseResult

TemporaryPlaceholderType