Struct ocl_core::types::structs::ImageFormat

pub struct ImageFormat {
    pub channel_order: ImageChannelOrder,
    pub channel_data_type: ImageChannelDataType,
}

Image format properties used by Image.

A structure that describes format properties of the image to be allocated. (from SDK)

Examples (from SDK)

To specify a normalized unsigned 8-bit / channel RGBA image: image_channel_order = CL_RGBA image_channel_data_type = CL_UNORM_INT8

image_channel_data_type values of CL_UNORM_SHORT_565, CL_UNORM_SHORT_555 and CL_UNORM_INT_101010 are special cases of packed image formats where the channels of each element are packed into a single unsigned short or unsigned int. For these special packed image formats, the channels are normally packed with the first channel in the most significant bits of the bitfield, and successive channels occupying progressively less significant locations. For CL_UNORM_SHORT_565, R is in bits 15:11, G is in bits 10:5 and B is in bits 4:0. For CL_UNORM_SHORT_555, bit 15 is undefined, R is in bits 14:10, G in bits 9:5 and B in bits 4:0. For CL_UNORM_INT_101010, bits 31:30 are undefined, R is in bits 29:20, G in bits 19:10 and B in bits 9:0. OpenCL implementations must maintain the minimum precision specified by the number of bits in image_channel_data_type. If the image format specified by image_channel_order, and image_channel_data_type cannot be supported by the OpenCL implementation, then the call to clCreateImage will return a NULL memory object.

Fields
channel_order: ImageChannelOrder channel_data_type: ImageChannelDataType
Methods

impl ImageFormat

pub fn new(
order: ImageChannelOrder,
data_type: ImageChannelDataType
) -> ImageFormat

pub fn new_rgba() -> ImageFormat

pub fn from_raw(fmt_raw: cl_image_format) -> ImageFormatParseResult

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn list_from_raw(
list_raw: Vec<cl_image_format>
) -> Vec<ImageFormatParseResult>

pub fn to_raw(&self) -> cl_image_format

pub fn new_raw() -> cl_image_format

pub fn pixel_bytes(&self) -> usize

Returns the size in bytes of a pixel using the format specified by this ImageFormat.

TODO: Add a special case for Depth & DepthStencil (https://www.khronos.org/registry/cl/sdk/2.0/docs/man/xhtml/cl_khr_gl_depth_images.html).

TODO: Validate combinations. TODO: Use core::get_image_info to check these with a test.