Struct ocl_core::types::structs::ContextProperties

pub struct ContextProperties { /* fields omitted */ }

Context properties list.

[MINIMALLY TESTED]

TODO: Check for duplicate property assignments.

Methods

impl ContextProperties

pub fn new() -> ContextProperties

Returns an empty new list of context properties

pub fn platform<P: Into<PlatformId>>(self, platform: P) -> ContextProperties

Specifies a platform (builder-style).

pub fn interop_user_sync(self, sync: bool) -> ContextProperties

Specifies whether the user is responsible for synchronization between OpenCL and other APIs (builder-style).

pub fn gl_context(self, gl_ctx: *mut c_void) -> ContextProperties

Specifies an OpenGL context handle (builder-style).

pub fn glx_display(self, glx_disp: *mut c_void) -> ContextProperties

Specifies a Display pointer for the GLX context (builder-style).

pub fn wgl_hdc(self, wgl_hdc: *mut c_void) -> ContextProperties

Specifies a Display pointer for the WGL HDC (builder-style).

pub fn cgl_sharegroup(self, gl_sharegroup: *mut c_void) -> ContextProperties

Specifies an OpenGL context CGL share group to associate the OpenCL context with (builder-style).

pub fn egl_display(self, egl_disp: *mut c_void) -> ContextProperties

Specifies a pointer for the EGL display (builder-style).

pub fn property_value(self, prop: ContextPropertyValue) -> ContextProperties

Pushes a ContextPropertyValue onto this list of properties (builder-style).

pub fn set_platform<P: Into<PlatformId>>(&mut self, platform: P)

Specifies a platform.

pub fn set_interop_user_sync(&mut self, sync: bool)

Specifies whether the user is responsible for synchronization between OpenCL and other APIs.

pub fn set_gl_context(&mut self, gl_ctx: *mut c_void)

Specifies an OpenGL context handle.

pub fn set_glx_display(&mut self, glx_disp: *mut c_void)

Specifies a Display pointer for the GLX context.

pub fn set_wgl_hdc(&mut self, wgl_hdc: *mut c_void)

Specifies a Display pointer for the WGL HDC.

pub fn set_cgl_sharegroup(&mut self, gl_sharegroup: *mut c_void)

Specifies an OpenGL context CGL share group to associate the OpenCL context with.

pub fn set_egl_display(&mut self, egl_disp: *mut c_void)

Specifies a pointer for the EGL display.

pub fn set_property_value(&mut self, prop: ContextPropertyValue)

Pushes a ContextPropertyValue onto this list of properties.

pub fn get_platform(&self) -> Option<PlatformId>

Returns a platform id or none.

pub fn contains_gl_context_or_sharegroup(&self) -> bool

Returns true if this set of context properties specifies any OpenGL context or sharegroup to associate with.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn to_raw(&self) -> Vec<isize>

Converts this list into a packed-word representation as specified here.

pub unsafe fn extract_property_from_raw(
property: ContextProperty,
raw_context_properties: &[isize]
) -> Option<ContextPropertyValue>

Returns a single context property value.

pub unsafe fn from_raw(
raw_context_properties: &[isize]
) -> OclCoreResult<ContextProperties>

Converts raw stuff into other stuff.